refactor: Fallback to workspace if workspace_default_members is unavailable
When cargo_metadata is invoked with a Cargo version < 1.71, which doesn't support workspace default members, the workspace_default_members field in cargo_metadata will be empty. Dereferencing this field will then cause a panic. To prevent partition_packages to panic when used with older Cargo versions, we check the presence of the field, and if it is missing, we default to the workspace instead.